src/loadgenerator/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
certifi==2018.11.29       # via requests
chardet==3.0.4            # via requests
click==7.0                # via flask
flask==1.0.2              # via locustio
gevent==1.4.0             # via locustio
greenlet==0.4.15          # via gevent
idna==2.8                 # via requests
itsdangerous==1.1.0       # via flask
jinja2==2.10              # via flask
locustio==0.9.0
markupsafe==1.1.0         # via jinja2
msgpack==0.6.1            # via locustio
pyzmq==17.0.0             # via locustio
requests==2.21.0          # via locustio
six==1.12.0               # via locustio
urllib3==1.24.1           # via requests
werkzeug==0.14.1          # via flask
